pyproject.toml
[project]
name = "ikkf"
version = "0.1.0"
description = "Short for 'I Know Kung Fu', instant context-aware skill loading for your agents"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["llm", "agent", "prompt", "skills", "ai", "anthropic"]
authors = [
    { name = "Matt Jenior", email = "mattjenior@gmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "semantic-router>=0.1.12",
    "litellm>=1.81.0",
]

[project.urls]
Documentation = "https://github.com/mjenior/IKKF/README.md"
Issues = "https://github.com/mjenior/IKKF/issues"
Source = "https://github.com/mjenior/IKKF"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ikkf"]

[tool.hatch.build.targets.sdist]
include = [
    "ikkf/**/*.py",
    "ikkf/py.typed",
    "ikkf/core/**/*.yaml",
    "ikkf/core/**/*.md",
    "README.md",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["ikkf"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
